+++ src/test/groovy/nebula/plugin/dependencylock/DependencyLockLauncherSpec.groovy
@@ -72,6 +72,25 @@ class DependencyLockLauncherSpec extends IntegrationSpec {
         standardOutput.contains 'com.google.guava:guava:14.0.1 -> 14.0'
     }
 
+    def 'override lock file is applied'() {
+        def dependenciesLock = new File(projectDir, 'dependencies.lock')
+        dependenciesLock << OLD_GUAVA_LOCK
+
+        def testOverride = new File(projectDir, 'test.override')
+        testOverride << GUAVA_LOCK
+
+        def gradleProperties = new File(projectDir, 'gradle.properties')
+        gradleProperties << 'dependencyLock.lockFile = \'test.override\''
+
+        buildFile << BUILD_GRADLE
+
+        when:
+        runTasksSuccessfully('dependencies')
+
+        then:
+        standardOutput.contains 'com.google.guava:guava:14.+ -> 14.0.1'
+    }
+
     def 'create lock'() {
         buildFile << BUILD_GRADLE